modify sort library to allow reusage
