+++ sort.c
@@ -1,30 +1,7 @@
+#include "sort.h"
 #include "types.h"
 #include "stat.h"
 #include "user.h"
-
-void swap(int *x, int *y)
-{
-  int temp = *x;
-  *x = *y;
-  *y = temp;
-}
-
-void selectionSort(int arr[], int n)
-{
-  int i, j, min_idx;
-
-  for (i = 0; i < n-1; i++)
-  {
-    min_idx = i;
-    for (j = i+1; j < n; j++)
-    {
-      if(arr[j] < arr[min_idx])
-        min_idx = j;
-    }
-      if(min_idx != i)
-      swap(&arr[min_idx], &arr[i]);
-  }
-}
 
 void sort(int argc, char *argv[])
 {
@@ -47,7 +24,7 @@ void sort(int argc, char *argv[])
   {
     printf(1, "%d ", arr[i]);
   }
-  
+
   printf(1, "\n");
 }
